pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "surrogate-rules"
version = "0.1.0"
description = "MIP-based surrogate rule extraction from Random Forests"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scipy>=1.10",
    "scikit-learn>=1.3",
    "gurobipy>=10.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "pymongo>=4.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "mongomock>=4.1",
    "ruff>=0.4",
    "mypy>=1.8",
]

[project.scripts]
run-experiment = "scripts.run_experiment:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
